Moved demos into dedicated folder
